separate is_translated_taxonomy check from elseif
Having is_translated_taxonomy in the elseif statement causes the if/else to move on the next item. For an archive page of a taxonomy that is NOT translated, the elseif for is_archive() is reached, and then ends up checking against the count of post_type posts in each language, rather than if there are posts related to that taxonomy. 

e.g. On mysite.com/non-translated-tax/the-tax-slug/, my site has German and Japanese languages and posts that are post_type 'post', I get hreflang links for mysite.com/de/non-translated-tax/the-tax-slug/, and mysite.com/ja/non-translated-tax/the-tax-slug/, even though these pages do not exist.

Moving the is_translated_taxonomy into a nested if() seems to fix this. This mimics the is_translated_post_type nested inside the is_post_type_archive check on lines ~140